Update month.rb
I haven't run it, but that should be a better way.   looks like you were shifting the array but not storing the values.  also treating the results as an array for longer will help and then just join the array when your done.  


on the first week I switched to shift so that your middle weeks doesn't print out first week again.    


I think you should fix your month array to also add the spaces to the end of the month to fill those spaces as well.   that way your week helpers wouldn't need to be any different.   you would just be joining with space and adding newlines.